Refactor code by moving them into separate files
